4.28. Refatorando a exclusão de cozinhas para usar domain services
* Refatorando a exclusão de cozinhas para usar domain services
* Utilizando Exceções do Spring ao trabalhar com Banco de Dados

+++ gorisfood-api/src/main/java/br/com/felipe/gorisfood/infrasctructure/repository/CozinhaRepositoryImpl.java
@@ -5,6 +5,7 @@ import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -37,13 +38,13 @@ public class CozinhaRepositoryImpl implements CozinhaRepository {
 	@Transactional
 	public void remover(Long id) {
 		Cozinha cozinha = this.buscar(id);
+		
+		if(cozinha == null) {
+			throw new EmptyResultDataAccessException(1);
+		}
+		
 		em.remove(cozinha);
 	}
 	
-	@Override
-	@Transactional
-	public void remover(Cozinha cozinha) {
-		em.remove(cozinha);
-	}
 
 }
